Fix guideListPages codec
