+++ .changeset/warm-dryers-dance.md
@@ -1,0 +1,5 @@
+---
+"nextjs-website": patch
+---
+
+Refactor guideListPages codec
